+++ CHANGES.rst
@@ -10,6 +10,10 @@
 - ``FileUpload`` objects now support the ``seekable()`` method on Python 3.
   Fixes `issue 44 <https://github.com/zopefoundation/zope.publisher/issue/44>`_.
 
+- Character set handling was rather comprehensively broken on Python 3.
+  It should be fixed now.  See `issue 41
+  <https://github.com/zopefoundation/zope.publisher/issue/41>`_.
+
 
 5.0.1 (2018-10-19)
 ==================